refactor connectors
