+++ backend/workflow_manager/endpoint/constants.py
@@ -8,7 +8,6 @@ class DBConnectionClass:
     SNOWFLAKE = "SnowflakeDB"
     BIGQUERY = "BigQuery"
     MSSQL = "MSSQL"
-    PostgreSQL = "PostgreSQL"
 
 
 class Snowflake:
