+++ unstract/connectors/src/unstract/connectors/databases/mariadb/mariadb.py
@@ -5,10 +5,11 @@ import pymysql
 from pymysql.connections import Connection
 
 from unstract.connectors.databases.mysql import MySQL
+from unstract.connectors.databases.mysql_handler import MysqlHandler
 from unstract.connectors.databases.unstract_db import UnstractDB
 
 
-class MariaDB(UnstractDB):
+class MariaDB(UnstractDB, MysqlHandler):
     def __init__(self, settings: dict[str, Any]):
         super().__init__("MariaDB")
 
@@ -74,3 +75,13 @@ class MariaDB(UnstractDB):
             Mysql and Mariadb share same SQL column type
         """
         return str(MySQL.sql_to_db_mapping(value=value))
+
+    def execute_query(
+        self, engine: Any, sql_query: str, sql_values: Any, **kwargs: Any
+    ) -> None:
+        MysqlHandler.execute_query(
+            engine=engine,
+            sql_query=sql_query,
+            sql_values=sql_values,
+            database=self.database,
+        )
